+++ src/tictactoe.h
@@ -20,7 +20,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #include <stdio.h>
 #include <stdlib.h>
 
-extern int_fast32_t leaf_nodes_evaluated;
+extern int_fast16_t leaf_nodes_evaluated;
 
 void computer_move(int_fast8_t* board,int_fast8_t side);
 int_fast8_t
